requirements.txt
Pyrogram
TgCrypto
requests
wget
speedtest-cli
